Updated Carousel; first implementation of touch swiping
