+++ package.json
@@ -38,6 +38,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.7.1",
+    "react-swipeable": "^7.0.0",
     "react-transition-group": "^4.4.5",
     "react-uuid": "^2.0.0",
     "telegram": "^2.15.2",
