+++ src/utils/references.ts
@@ -6,16 +6,16 @@ export const REFERENCES = [
     bkg_head: "#27303D",
   },
   {
-    handle: "BuxtonMatthew",
-    subhandle: "Founder @ Mystic Games",
-    text: "A platform like TheCoreLoop provides a trusted source for legitimate games to boost their reputation in the space. Moreover, it further allows us to scope out potential collaborators in the gaming sector.",
-    bkg_head: "#45303F",
-  },
-  {
     handle: "quichenpatel",
     subhandle: "Investments @ BITKRAFT",
     text: "Info on good web3 games is extremely scattered and frankly, there are many low quality games that crowd out quality. As an investor in web3 gaming, what TheCoreLoop is building would be awesome",
     bkg_head: "#263670",
+  },
+  {
+    handle: "BuxtonMatthew",
+    subhandle: "Founder @ Mystic Games",
+    text: "A platform like TheCoreLoop provides a trusted source for legitimate games to boost their reputation in the space. Moreover, it further allows us to scope out potential collaborators in the gaming sector.",
+    bkg_head: "#45303F",
   },
   {
     handle: "snake_chain",
@@ -24,15 +24,15 @@ export const REFERENCES = [
     bkg_head: "#323A53",
   },
   {
-    handle: "shloked_",
-    subhandle: "Partnerships Lead @ Glip.gg",
-    text: "What would matter to me is learning about new high quality projects that isn’t often covered compared to the established ones. I definitely see a lot of value here at TheCoreLoop.",
-    bkg_head: "#5D604A",
-  },
-  {
     handle: "__sugs",
     subhandle: "Investments @ Newman Capital",
     text: "As a VC investor, I’d like a discovery tool for finding new projects based on whatever categories I'm searching for (ex. fps, mmorpg, latest trending, building in stealth) - this could give me a sense of what's out there, who's seeking funding, etc",
     bkg_head: "#2E434D",
+  },
+  {
+    handle: "shloked_",
+    subhandle: "Partnerships Lead @ Glip.gg",
+    text: "What would matter to me is learning about new high quality projects that isn’t often covered compared to the established ones. I definitely see a lot of value here at TheCoreLoop.",
+    bkg_head: "#5D604A",
   },
 ];
